SSR in Progress:
/two currently renders from the server but / is delivering the empty template instead of the server...

A few other in progress items included attempts to get modernizr + foundation loaded in App.vue to no avail. We will likely have to follow the errors mocking out globals as we find them to make this work. I think a better route would be to Webpack Provide or Expose these only in the client.

Will work on integrating SSR into dev mode soon to make it easier to work.